Renaming workbox-runtime-caching to workbox-strategies
