+++ demos/functions/views/demo/workbox-strategies.hbs
@@ -26,7 +26,7 @@
 </section>
 
 <script>
-  navigator.serviceWorker.register('/demo/workbox-runtime-caching/sw.js');
+  navigator.serviceWorker.register('/demo/workbox-strategies/sw.js');
 
   const cacheOnlyEmpty = document.querySelector('.cache-only-empty-cache');
   cacheOnlyEmpty.addEventListener('click', () => {
