+++ demos/functions/views/demo/workbox-strategies/sw.hbs
@@ -2,31 +2,31 @@
 
 self.addEventListener('fetch', (event) => {
   switch (new URL(event.request.url).pathname) {
-    case '/demo/workbox-runtime-caching/cache-only-empty-cache.txt':
+    case '/demo/workbox-strategies/cache-only-empty-cache.txt':
       const cacheOnlyEmpty = new workbox.strategies.CacheOnly();
       event.respondWith(cacheOnlyEmpty.handle({event}));
       break;
-    case '/demo/workbox-runtime-caching/cache-only-populated-cache':
+    case '/demo/workbox-strategies/cache-only-populated-cache':
       const cacheOnlyPopulated = new workbox.strategies.CacheOnly();
       event.respondWith(cacheOnlyPopulated.handle({event}));
       break;
-    case '/demo/workbox-runtime-caching/cache-first.txt':
+    case '/demo/workbox-strategies/cache-first.txt':
       const cacheFirst = new workbox.strategies.CacheFirst();
       event.respondWith(cacheFirst.handle({event}));
       break;
-    case '/demo/workbox-runtime-caching/network-only.txt':
+    case '/demo/workbox-strategies/network-only.txt':
       const networkOnly = new workbox.strategies.NetworkOnly();
       event.respondWith(networkOnly.handle({event}));
       break;
-    case '/demo/workbox-runtime-caching/network-first.txt':
+    case '/demo/workbox-strategies/network-first.txt':
       const networkFirst = new workbox.strategies.NetworkFirst();
       event.respondWith(networkFirst.handle({event}));
       break;
-    case '/demo/workbox-runtime-caching/network-first-404.txt':
+    case '/demo/workbox-strategies/network-first-404.txt':
       const networkFirstInvalid = new workbox.strategies.NetworkFirst();
       event.respondWith(networkFirstInvalid.handle({event}));
       break;
-    case '/demo/workbox-runtime-caching/stale-while-revalidate.txt':
+    case '/demo/workbox-strategies/stale-while-revalidate.txt':
       const staleWhileRevalidate = new workbox.strategies.StaleWhileRevalidate();
       event.respondWith(staleWhileRevalidate.handle({event}));
       break;
@@ -44,7 +44,7 @@ self.addEventListener('install', (event) => {
     caches.open(workbox.core.cacheNames.runtime)
     .then((cache) => {
       return cache.put(
-        new Request('/demo/workbox-runtime-caching/cache-only-populated-cache'),
+        new Request('/demo/workbox-strategies/cache-only-populated-cache'),
         new Response('Hello from the populated cache.')
       );
     })
