+++ demos/public/demo/workbox-strategies/cache-first.txt
