+++ demos/public/demo/workbox-strategies/network-first.txt
