+++ packages/workbox-google-analytics/package.json
@@ -31,6 +31,6 @@
     "workbox-background-sync": "^3.0.0-alpha.24",
     "workbox-core": "^3.0.0-alpha.24",
     "workbox-routing": "^3.0.0-alpha.24",
-    "workbox-runtime-caching": "^3.0.0-alpha.24"
+    "workbox-strategies": "^3.0.0-alpha.24"
   }
 }
