--- packages/workbox-runtime-caching/_version.mjs
@@ -1,1 +1,0 @@
-try{self.workbox.v['workbox:runtime-caching:3.0.0-alpha.24']=1;}catch(e){} // eslint-disable-line
