+++ packages/workbox-strategies/_version.mjs
@@ -1,0 +1,1 @@
+try{self.workbox.v['workbox:strategies:3.0.0-alpha.24']=1;}catch(e){} // eslint-disable-line
