+++ packages/workbox-strategies/plugins/cacheOkAndOpaquePlugin.mjs
