+++ packages/workbox-strategies/utils/pluginBuilder.mjs
