+++ test/workbox-build/node/entry-points/generate-sw.js
@@ -165,10 +165,10 @@ describe(`[workbox-build] entry-points/generate-sw.js (End to End)`, function() 
         'workbox-routing.dev.js.map',
         'workbox-routing.prod.js',
         'workbox-routing.prod.js.map',
-        'workbox-runtime-caching.dev.js',
-        'workbox-runtime-caching.dev.js.map',
-        'workbox-runtime-caching.prod.js',
-        'workbox-runtime-caching.prod.js.map',
+        'workbox-strategies.dev.js',
+        'workbox-strategies.dev.js.map',
+        'workbox-strategies.prod.js',
+        'workbox-strategies.prod.js.map',
         'workbox-sw.js',
         'workbox-sw.js.map',
       ]);
