+++ test/workbox-runtime-caching/node/test-CacheFirst.mjs
@@ -18,9 +18,9 @@ import {expect} from 'chai';
 import {_private} from '../../../packages/workbox-core/index.mjs';
 import {compareResponses} from '../utils/response-comparisons.mjs';
 
-import {CacheFirst} from '../../../packages/workbox-runtime-caching/CacheFirst.mjs';
+import {CacheFirst} from '../../../packages/workbox-strategies/CacheFirst.mjs';
 
-describe(`[workbox-runtime-caching] CacheFirst`, function() {
+describe(`[workbox-strategies] CacheFirst`, function() {
   let sandbox = sinon.sandbox.create();
 
   beforeEach(async function() {
