+++ test/workbox-runtime-caching/node/test-NetworkOnly.mjs
@@ -18,9 +18,9 @@ import {expect} from 'chai';
 
 import {_private} from '../../../packages/workbox-core/index.mjs';
 
-import {NetworkOnly} from '../../../packages/workbox-runtime-caching/NetworkOnly.mjs';
+import {NetworkOnly} from '../../../packages/workbox-strategies/NetworkOnly.mjs';
 
-describe(`[workbox-runtime-caching] NetworkOnly`, function() {
+describe(`[workbox-strategies] NetworkOnly`, function() {
   let sandbox = sinon.sandbox.create();
 
   beforeEach(async function() {
